v4 architecture: iterative memory refinement via causal scan
Replace v1/v2 architecture docs with v4 design — cortical columns with
iterative causal prefix scan over structured memory state (PM/EM/WM/PCM).

Key idea: separate non-affine decisions (Hebbian, episodic) from affine
state updates (S = A*S + b), enabling parallel scan while preserving
bio-inspired memory mechanisms. R iterative passes give depth.

Also includes WM scatter dtype fix (bf16/fp32 mismatch on CUDA).
Old architecture preserved on v1-legacy branch.

+++ src/model/working_memory.py
@@ -381,8 +381,8 @@ class WorkingMemory(nn.Module, StateMixin):
         # --- Differentiable cache update (write all P tokens at once) ---
         # P <= W is guaranteed by config validation, so write_pos is unique per stream.
         scatter_idx = write_pos.unsqueeze(-1).expand(-1, -1, self.D_wm)  # [BS, P, D_wm]
-        self.wm_K = self.wm_K.scatter(1, scatter_idx, k_all)
-        self.wm_V = self.wm_V.scatter(1, scatter_idx, v_all)
+        self.wm_K = self.wm_K.scatter(1, scatter_idx, k_all.to(self.wm_K.dtype))
+        self.wm_V = self.wm_V.scatter(1, scatter_idx, v_all.to(self.wm_V.dtype))
 
         # --- Update wm_valid and advance wm_ptr ---
         new_valid = self.wm_valid.clone()
